requirements.txt
alembic==1.18.4
anyio==4.13.0
blinker==1.9.0
certifi==2026.4.22
click==8.3.3
colorama==0.4.6
Flask==3.1.3
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
greenlet==3.5.0
gunicorn==25.3.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.13
iniconfig==2.3.0
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.12
mangum==0.21.0
MarkupSafe==3.0.3
packaging==26.2
pluggy==1.6.0
psycopg2-binary==2.9.12
Pygments==2.20.0
pytest==9.0.3
python-dotenv==1.2.2
redis==7.4.0
SQLAlchemy==2.0.49
typing_extensions==4.15.0
upstash-redis==1.7.0
Werkzeug==3.1.8
